+++ libraries/base/changelog.md
@@ -18,6 +18,11 @@
 
   * `Data.Type.Coercion` now provides `gcoerceWith` (#12493)
 
+  * New methods `liftReadList(2)` and `liftReadListPrec(2)` in the
+    `Read1`/`Read2` classes that are defined in terms of `ReadPrec` instead of
+    `ReadS`, as well as related combinators, have been added to
+    `Data.Functor.Classes` (#12358)
+
 ## 4.9.0.0  *May 2016*
 
   * Bundled with GHC 8.0